requirements.txt
fastapi
uvicorn
jinja2
python-multipart
bcrypt
aiogram
requests
asyncio
python-dotenv